feat: push pin tar files
